refactor: remove circular dependencies in dev tools (#16261)
