+++ vaadin-dev-server/frontend/styles.ts
@@ -1,0 +1,27 @@
+import { css } from 'lit';
+
+export const popupStyles = css`
+  .popup {
+    width: auto;
+    position: fixed;
+    background-color: var(--dev-tools-background-color-active-blurred);
+    color: var(--dev-tools-text-color-primary);
+    padding: 0.1875rem 0.75rem 0.1875rem 1rem;
+    background-clip: padding-box;
+    border-radius: var(--dev-tools-border-radius);
+    overflow: hidden;
+    margin: 0.5rem;
+    width: 30rem;
+    max-width: calc(100% - 1rem);
+    max-height: calc(100vh - 1rem);
+    flex-shrink: 1;
+    background-color: var(--dev-tools-background-color-active);
+    color: var(--dev-tools-text-color);
+    transition: var(--dev-tools-transition-duration);
+    transform-origin: bottom right;
+    display: flex;
+    flex-direction: column;
+    box-shadow: var(--dev-tools-box-shadow);
+    outline: none;
+  }
+`;
